requirements.txt
boto==2.49.0
boto3==1.11.12
botocore==1.14.12
certifi==2018.8.24
chardet==3.0.4
diff-match-patch==20121119
dj-database-url==0.5.0
Django==2.0.5
django-countries==5.5
django-crispy-forms==1.7.2
django-filter==2.2.0
django-import-export==1.1.0
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
et-xmlfile==1.0.1
gmplot==1.2.0
gunicorn==20.0.4
idna==2.7
jdcal==1.4
jmespath==0.9.4
Markdown==3.1.1
odfpy==1.3.6
openpyxl==2.5.8
Pillow==5.3.0
postgis==1.0.4
psycopg2==2.7.5
python-dateutil==2.8.1
pytz==2018.5
PyYAML==3.13
requests==2.22.0
s3transfer==0.3.3
six==1.14.0
stripe==2.42.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.23
virtualenv==16.0.0
xlrd==1.1.0
xlwt==1.3.0
